requirements.txt
aiohttp==3.7.3
aioice==0.6.18
aiortc==1.0.0
async-timeout==3.0.1
attrs==20.3.0
av==8.0.2
cffi==1.14.4
chardet==3.0.4
crc32c==2.2
cryptography==3.3.1
idna==2.10
multidict==5.1.0
netifaces==0.10.9
numpy==1.19.4
opencv-contrib-python==4.4.0.46
opencv-python==4.4.0.46
Pillow==8.0.1
pycparser==2.20
pyee==8.1.0
pylibsrtp==0.6.7
six==1.15.0
typing-extensions==3.7.4.3
yarl==1.6.3
